+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.1.4
+
+* Bug fix (#41) that `github.get_contents` action would be failed when decode parameter is set,
+  and fix encoding processing problem in `github.create_file` and `github.update_file` actions.
+
 ## 2.1.3
 
 * Fix `update_branch_protection` action: dismissal users and teams can now be null in GitHub's API response.